refactor: botfuel-nlu uses Botfuel Trainer
BREAKING CHANGE: the nlu botfuel (default) will use Botfuel Trainer. Bots must be migrated to use Botfuel Trainer. No more intents folder and intents must be created inside the Botfuel Trainer.
